App icons: promote to its own Settings section (#407)
Move the Icon Selector out of the Application section into its own left-bar
section, so the order is: Application, App Icons, then the app-specific sections.

- New AppIconsViewModel + AppIconsView (the selector + commands moved here).
- SettingsWindowViewModel adds it as the second fixed section, after Application.
- AppSettingsViewModel/View reverted to their pre-icon state.
