+++ Jellyfin2Samsung-CrossOS/Views/AppIconsView.axaml
@@ -1,0 +1,61 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:viewModels="clr-namespace:Apps2Samsung.ViewModels"
+             xmlns:models="clr-namespace:Apps2Samsung.Models"
+             xmlns:fa="clr-namespace:FluentAvalonia.UI.Controls;assembly=FluentAvalonia"
+             mc:Ignorable="d"
+             x:Class="Apps2Samsung.Views.AppIconsView"
+             x:DataType="viewModels:AppIconsViewModel">
+
+	<ScrollViewer VerticalScrollBarVisibility="Auto" Padding="0,12,20,0">
+		<StackPanel Spacing="12">
+
+			<TextBlock Text="{Binding LblAppIcons}"
+					   FontSize="15"
+					   FontWeight="SemiBold"
+					   Foreground="{DynamicResource SystemControlForegroundBaseHighBrush}"/>
+			<TextBlock Text="{Binding LblAppIconsHint}"
+					   FontSize="12"
+					   Foreground="{DynamicResource SystemControlForegroundBaseMediumBrush}"
+					   TextWrapping="Wrap"/>
+
+			<ComboBox ItemsSource="{Binding AppIcons}"
+					  SelectedItem="{Binding SelectedAppIcon}"
+					  HorizontalAlignment="Left"
+					  MinWidth="320">
+				<ComboBox.ItemTemplate>
+					<DataTemplate x:DataType="models:AppIconEntry">
+						<TextBlock Text="{Binding DisplayName}"/>
+					</DataTemplate>
+				</ComboBox.ItemTemplate>
+			</ComboBox>
+
+			<Grid ColumnDefinitions="*,Auto" ColumnSpacing="8" IsVisible="{Binding HasSelectedAppIcon}">
+				<TextBlock Grid.Column="0"
+						   Text="{Binding SelectedAppIcon.Summary}"
+						   VerticalAlignment="Center"
+						   TextTrimming="CharacterEllipsis"
+						   Foreground="{DynamicResource SystemControlForegroundBaseMediumBrush}"/>
+				<StackPanel Grid.Column="1" Orientation="Horizontal" Spacing="4">
+					<Button Classes="secondary" Padding="10,6"
+							IsVisible="{Binding SelectedAppIcon.HasOblong}"
+							Command="{Binding UseOblongIconCommand}">
+						<TextBlock Text="{Binding LblIconOblong}"/>
+					</Button>
+					<Button Classes="secondary" Padding="10,6"
+							Command="{Binding BrowseCustomIconCommand}">
+						<TextBlock Text="{Binding LblIconCustom}"/>
+					</Button>
+					<Button Classes="secondary" Padding="10,6"
+							IsVisible="{Binding !SelectedAppIcon.IsDefault}"
+							Command="{Binding ResetIconCommand}">
+						<TextBlock Text="{Binding LblIconReset}"/>
+					</Button>
+				</StackPanel>
+			</Grid>
+
+		</StackPanel>
+	</ScrollViewer>
+</UserControl>
