+++ Jellyfin2Samsung-CrossOS/Views/AppIconsView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace Apps2Samsung.Views;
+
+public partial class AppIconsView : UserControl
+{
+    public AppIconsView()
+    {
+        InitializeComponent();
+    }
+}
